+++ src/DependentTypes/DependentTypes.fs
@@ -23,65 +23,22 @@ module Helpers =
             (^S: (static member Extract: ^S -> ^T option ) x') )
             |> Option.flatten).Value
      
-//module DependentTypes = Helpers   //aliasing not visible externally
-/// deprecated in favor of Helpers
-module DependentTypes =
-    /// <summary>
-    /// Create instance of dependent type.
-    /// </summary>
-    /// <remarks>
-    /// Deprecated. Use Helpers.someValue.
-    /// </remarks>
-    let inline mkDependentType (x: ^S) : ^T = 
-        (^T: (static member Create: ^S -> ^T) x)
-
-    /// <summary>
-    /// Retrieves 'T2 (base type) value.
-    /// </summary>
-    /// <remarks>
-    /// Deprecated. Use Helpers.someValue.
-    /// </remarks>
-    let inline extract (x:^S) = 
-        (^S: (static member Extract: ^S -> ^T) x)
-
-    /// <summary>
-    /// Conversion of 'T2 (base type) element value to compatible dependent type.
-    /// </summary>
-    /// <remarks>
-    /// Deprecated. Use Helpers.someValue.
-    /// </remarks>
-    let inline convertTo (x: ^S) : ^T = 
-        (^T: (static member ConvertTo: ^S -> ^T) x)
-
-    /// <summary>
-    /// Retrieves the 'T2 (base type) element from a DependentType option.
-    /// </summary>
-    /// <remarks>
-    /// Deprecated. Use Helpers.someValue.
-    /// </remarks>
-    /// <exception cref="System.NullReferenceException">Thrown when None.</exception>
-    let inline someValue (x : ^S Option) =
-        (x |> Option.map (fun x' ->
-            (^S: (static member Extract: ^S -> ^T option ) x') )
-            |> Option.flatten).Value
-
-
 open Helpers
 
 [<Class>]
 /// Construction / validation type for DependentType 'T -> 'T2 
-type PiType<'Config, 'T, 'T2> (config: 'Config, pi: 'Config -> 'T -> 'T2) =
+type Pi<'Config, 'T, 'T2> (config: 'Config, pi: 'Config -> 'T -> 'T2) =
     member __.Create x  = pi config x
 
 
 [<Class>]
 /// Construction / validation type for DependentPair 'T -> 'T * 'T2
-type SigmaType<'Config, 'T, 'T2> (config: 'Config, pi: 'Config -> 'T -> 'T2) =
+type Sigma<'Config, 'T, 'T2> (config: 'Config, pi: 'Config -> 'T -> 'T2) =
     member __.Create(x:'T) : 'T * 'T2 = x, (pi config x)
 
 /// 'T -> 'T2 dependent type
-type DependentType<'PiType, 'Config, 'T, 'T2 when 'PiType :> PiType<'Config, 'T, 'T2>  
-                                              and  'PiType : (new: unit -> 'PiType)> =
+type DependentType<'Pi, 'Config, 'T, 'T2 when 'Pi :> Pi<'Config, 'T, 'T2>  
+                                              and  'Pi : (new: unit -> 'Pi)> =
     DependentType of 'T2
     with 
         /// 'T2 (base type) element value.
@@ -93,19 +50,19 @@ type DependentType<'PiType, 'Config, 'T, 'T2 when 'PiType :> PiType<'Config, 'T,
         override __.ToString() = __.Value.ToString()     
         
         /// Retrieve 'T2 (base type) element value.
-        static member Extract  (x : DependentType<'PiType, 'Config, 'T, 'T2> ) = 
+        static member Extract  (x : DependentType<'Pi, 'Config, 'T, 'T2> ) = 
             let (DependentType t2) = x
             t2
 
         /// Create instance of dependent type.
-        static member Create x : DependentType<'PiType, 'Config, 'T, 'T2> =
-            (new 'PiType()).Create x
+        static member Create x : DependentType<'Pi, 'Config, 'T, 'T2> =
+            (new 'Pi()).Create x
             |> DependentType
 
         /// Create instance of DependentType option.
         /// If the 'T2 (base type) is option, lifts Some/None of base type element to DependentType option.
-        static member TryCreate x : DependentType<'PiType, 'Config, 'T, 'T2> Option =
-            let piResult = (new 'PiType()).Create x
+        static member TryCreate x : DependentType<'Pi, 'Config, 'T, 'T2> Option =
+            let piResult = (new 'Pi()).Create x
 
             if typedefof<'T2> = typedefof<_ option> then
                 if isNull (piResult :> obj) then
@@ -117,7 +74,7 @@ type DependentType<'PiType, 'Config, 'T, 'T2 when 'PiType :> PiType<'Config, 'T,
 
         /// Create instance of DependentType option.
         /// If the 'T2 (base type) is option, lifts Some/None of base type element to DependentType option.
-        static member TryCreate (x : 'T Option) : DependentType<'PiType, 'Config, 'T, 'T2> Option =
+        static member TryCreate (x : 'T Option) : DependentType<'Pi, 'Config, 'T, 'T2> Option =
             match x with
             | Some t -> 
                 DependentType.TryCreate t
@@ -131,8 +88,8 @@ type DependentType<'PiType, 'Config, 'T, 'T2 when 'PiType :> PiType<'Config, 'T,
             mkDependentType t2  
 
 /// 'T -> 'T * 'T2 dependent pair
-type DependentPair<'SigmaType, 'Config, 'T, 'T2 when 'SigmaType :> SigmaType<'Config, 'T, 'T2>  
-                                                 and 'SigmaType : (new: unit -> 'SigmaType)> =
+type DependentPair<'Sigma, 'Config, 'T, 'T2 when 'Sigma :> Sigma<'Config, 'T, 'T2>  
+                                                 and 'Sigma : (new: unit -> 'Sigma)> =
      DependentPair of 'T * 'T2
      with 
         /// Pair of 'T1 element and 'T2 (base type) element value.
@@ -141,6 +98,6 @@ type DependentPair<'SigmaType, 'Config, 'T, 'T2 when 'SigmaType :> SigmaType<'Co
             s, s2
 
         /// Create instance of dependent pair type.
-        static member Create(x:'T) : DependentPair<'SigmaType, 'Config, 'T, 'T2> =
-            (new 'SigmaType()).Create x
+        static member Create(x:'T) : DependentPair<'Sigma, 'Config, 'T, 'T2> =
+            (new 'Sigma()).Create x
             |> DependentPair