+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
-#### 0.1.1 - unreleased, 2018
+#### 0.2.0 - unreleased, 2018
+* breaking change
+* module DependentTypes.DependentTypes renamed DependentTypes.Helpers
+* PiType renamed to Pi
+* SigmaType renamed to Sigma
 * more efficient option type lift
-* deprecate module DependentTypes.DependentTypes in favor of DependentTypes.Helpers
 
 #### 0.1.0 - September 25, 2018
 * breaking change